requirements.txt
numba==0.57.1
Pillow==10.0.0